+++ oneEngine-core/source/core/system/io/CMappedBinaryFile.h
@@ -21,14 +21,14 @@ public:
 
 	// Sets the buffer to 0, then reads in as much data as possible.
 	// Does not expand the file.
-	uint32_t	ReadBuffer ( void* n_buffer, const uint32_t n_buffer_size );
+	size_t		ReadBuffer ( void* n_buffer, const size_t n_buffer_size );
 	// Writes the buffer to the file, overwriting data.
 	// Expands file as necessary.
-	uint32_t	WriteBuffer ( const void* n_buffer, const uint32_t n_buffer_size );
+	size_t		WriteBuffer ( const void* n_buffer, const size_t n_buffer_size );
 
 	// Seeks to the given location.
 	// Will expand the file as necessary.
-	void		SeekTo ( const uint32_t n_pos );
+	void		SeekTo ( const size_t n_pos );
 
 	// Returns a stream for manual I/O. May return NULL if manual I/O is not supported.
 	FILE*		GetStream ( void );
@@ -40,8 +40,8 @@ private:
 	arstring<256>	m_filename;
 
 private:
-	uint32_t	TellPos ( void );
-	uint32_t	TellSize( void );
+	size_t		TellPos ( void );
+	size_t		TellSize( void );
 
 };
